% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPIMM.R
\name{addPIMM}
\alias{addPIMM}
\title{add the 'percent immature mRNA' estimates to a SingleCellExperiment}
\usage{
addPIMM(exsce, intsce)
}
\arguments{
\item{exsce}{SingleCellExperiment with exon counts}

\item{intsce}{SingleCellExperiment with intron counts}
}
\value{
SingleCellExperiment exsce with an additional 
colData column PIMM with the average value over all genes for
each cell of the quantity (intronCount/(intronCount+exonCount))
}
\description{
add the 'percent immature mRNA' estimates to a SingleCellExperiment
}
