% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recorder.R
\name{biplots}
\alias{biplots}
\title{produce biplots of top genes for selections}
\usage{
biplots(sce, sels, which = 1, ntopgenes = 6)
}
\arguments{
\item{sce}{a SingleCellExperiment instance to which pcmpApp was applied}

\item{sels}{an instance of PcmpSels}

\item{which}{a numeric(4) collection telling which of the selections occupy a 2 x 2 biplot array}

\item{ntopgenes}{numeric(1) number of (most differentially-expressed, selection vs all others) genes to use in PCA for selected samples}
}
\description{
produce biplots of top genes for selections
}
\examples{
op = par(no.readonly=TRUE)
par(mfrow=c(2,2))
for (i in 1:4) biplots(pcmp::sce300xx, pcmp::acc4vis2, which=i)
par(op)
}
