% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\name{biplotSel}
\alias{biplotSel}
\title{simple biplot display of transformed expression data for a subset of cluster discriminating genes}
\usage{
biplotSel(sce, selnum = 1, assaytx = function(x) log(x + 1),
  ntop = 6, ...)
}
\arguments{
\item{sce}{a SingleCellExperiment as produced by pcmpApp}

\item{selnum}{numeric(1) the cluster to display}

\item{assaytx}{a function to be applied to the assay data}

\item{ntop}{numeric(1) number of genes to use in PCA}

\item{\dots}{additional parameters to biplot.prcomp, which should
exclude xlabs, expand, cex, and col which are hard coded.}
}
\description{
simple biplot display of transformed expression data for a subset of cluster discriminating genes
}
\examples{
biplotSel(pcmp::vigAccum, selnum=4, ntop=10)
}
