% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recorder.R
\docType{methods}
\name{cellSets}
\alias{cellSets}
\alias{cellSets,PcmpSels-method}
\title{extract list of cells selected}
\usage{
cellSets(x)

\S4method{cellSets}{PcmpSels}(x)
}
\arguments{
\item{x}{instance of PcmpSels}
}
\description{
extract list of cells selected
}
