% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\name{discreteColdVars}
\alias{discreteColdVars}
\title{find discrete variables in colData}
\usage{
discreteColdVars(se, maxnuv = 25)
}
\arguments{
\item{se}{instance of SummarizedExperiment; SingleCellExperiment also works}

\item{maxnuv}{numeric(1) largest number of unique values that may be present in a variable that will be regarded as 'discrete'}
}
\value{
character vector of variable names satisfying discreteness condition
}
\description{
find discrete variables in colData
}
\examples{
discreteColdVars  # need example se
}
