% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recorder.R
\docType{methods}
\name{geneTable}
\alias{geneTable}
\alias{geneTable,PcmpSels-method}
\title{extract limma topTable}
\usage{
geneTable(x)

\S4method{geneTable}{PcmpSels}(x)
}
\arguments{
\item{x}{instance of PcmpSels}
}
\description{
extract limma topTable
}
