% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{immgenAnno}
\alias{immgenAnno}
\title{annotation of immgen mouse immune cell resource}
\format{data.frame}
\source{
\url{https://gist.github.com/nachocab/3d9f374e0ade031c475a}
}
\usage{
immgenAnno
}
\description{
annotation of immgen mouse immune cell resource
}
\examples{
head(immgenAnno)
}
\keyword{datasets}
