% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recorder.R
\docType{class}
\name{PcmpSels}
\alias{PcmpSels}
\alias{PcmpSels-class}
\title{organize record of selections in pcmp}
\usage{
PcmpSels(obj)
}
\arguments{
\item{obj}{a list with elements 'cells' and 'limmaTab'}
}
\description{
organize record of selections in pcmp

construct pcmp selections object
}
\section{Slots}{

\describe{
\item{\code{cellSets}}{list() of cell identifiers for each selection}

\item{\code{geneTable}}{a single data.frame with a variable `sel` indicating
the selections to which rows belong}
}}

