% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\name{plotSelMap}
\alias{plotSelMap}
\title{plot selection map from output of pcmpApp}
\usage{
plotSelMap(esce, proj, dim1 = 1, dim2 = 2, alpha = 0.5)
}
\arguments{
\item{esce}{an extended SingleCellExperiment, with metadata extended using pcmpApp}

\item{proj}{character(1) name of a reducedDims component of esce}

\item{dim1}{numeric(1) dimension to use as x axis}

\item{dim2}{numeric(1) dimension to use as y axis}

\item{alpha}{numeric(1) defaults to .5 for transparency used in geom_point}
}
\description{
plot selection map from output of pcmpApp
}
