% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recorder.R
\name{replay}
\alias{replay}
\title{replay a selection by plotting the selected cells}
\usage{
replay(sce, sels, whichProj, dim1 = 1, dim2 = 2)
}
\arguments{
\item{sce}{a SingleCellExperiment instance to which pcmpApp was applied}

\item{sels}{an instance of PcmpSels}

\item{whichProj}{character(1) name of reducedDims list element to use}

\item{dim1}{numeric(1) x axis of replay}

\item{dim2}{numeric(1) y axis of replay}
}
\description{
replay a selection by plotting the selected cells
}
\examples{
replay(pcmp::sce300xx, pcmp::acc4vis2, "TS", 3, 4)
}
