% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce300xx}
\alias{sce300xx}
\title{demonstration SingleCellExperiment}
\format{SingleCellExperiment}
\source{
see metadata component
}
\usage{
sce300xx
}
\description{
demonstration SingleCellExperiment
}
\note{
Developed by stratified random sampling, with strata
defined by donor (3 levels) and brain region (2 levels, ACC and VIS).
Retains 300 cells chosen randomly within each stratum.
}
\examples{
suppressPackageStartupMessages({
 requireNamespace("SingleCellExperiment")
})
pcmp::sce300xx
}
\keyword{datasets}
