% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\name{stratsampInds}
\alias{stratsampInds}
\title{create vector of indices constituting a stratified sample using a discrete variable to partition strata}
\usage{
stratsampInds(strata, nperstrat = 300)
}
\arguments{
\item{strata}{atomic vector}

\item{nperstrat}{size of sample to be retrieved}
}
\value{
numeric vector of indices
}
\description{
create vector of indices constituting a stratified sample using a discrete variable to partition strata
}
\note{
fails if any stratum has size less than 'nperstrat'
}
\examples{
mytags = c(1,1,1,2,2,3,1,1,1,2,2,3,3,3,3,2,2,2)
mydat = 1:length(mytags)
names(mydat) = mytags
inds = stratsampInds(mytags, nperstrat=2)
mydat[inds]
}
