% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vigAccum}
\alias{vigAccum}
\title{collection of selections from sce300xx, corresponding to vignette displays}
\format{PcmpSels instance}
\usage{
vigAccum
}
\description{
collection of selections from sce300xx, corresponding to vignette displays
}
\examples{
data(vigAccum)
head(geneTable(vigAccum),3)
}
\keyword{datasets}
