% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acc4vis2}
\alias{acc4vis2}
\title{collection of 6 selections from reduced sce300xx}
\format{SingleCellExperiment instance}
\usage{
acc4vis2
}
\description{
collection of 6 selections from reduced sce300xx
}
\note{
gene set used for limmaTabs formed by requiring MAD over samples > 0
}
\examples{
data(acc4vis2)
plotSelMap(acc4vis2, "TSNE2")
}
\keyword{datasets}
