% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acc4vis2}
\alias{acc4vis2}
\title{collection of 6 selections from sce300xx}
\format{PcmpSels instance}
\usage{
acc4vis2
}
\description{
collection of 6 selections from sce300xx
}
\examples{
data(acc4vis2)
data(sce300xx)
rd = reducedDims(sce300xx)
plot(rd$TSNE2[,1], rd$TSNE2[,2])
inds1 = acc4vis2@cellSets[[1]]
points(rd$TSNE2[ inds1, 1], rd$TSNE2[ inds1, 2], pch=19, col="red")
}
\keyword{datasets}
