% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\name{addProjections}
\alias{addProjections}
\title{add a series of projectors to transformed assay data
in a SingleCellExperiment}
\usage{
addProjections(sce, ncomp = 4, tx = function(x) log(x + 1), projectors,
  ...)
}
\arguments{
\item{sce}{SingleCellExperiment instance}

\item{ncomp}{numeric(1) number of components per projection}

\item{tx}{function that will transform assay quantifications,
defaults to `function(x) log(x+1)`}

\item{projectors}{a list of list of functions like that created}

\item{\dots}{not used
by `defaultProjectors()`}
}
\description{
add a series of projectors to transformed assay data
in a SingleCellExperiment
}
