% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recorder.R
\name{biplotSel}
\alias{biplotSel}
\title{produce biplot of top genes for a selection}
\usage{
biplotSel(sce, sels, which = 1, ntopgenes = 6)
}
\arguments{
\item{sce}{a SingleCellExperiment instance to which pcmpApp was applied}

\item{sels}{an instance of PcmpSels}

\item{which}{a numeric(1) telling which of the 
selections is used to form a biplot.  The 
order used is that of `geneSets(sels)`}

\item{ntopgenes}{numeric(1) number of 
(most differentially-expressed, selection vs all others) 
genes to use in PCA for selected samples; the
assay data are transformed as log(assay+1) before PCA is carried out}
}
\description{
produce biplot of top genes for a selection
}
\examples{
op = par(no.readonly=TRUE)
par(mfrow=c(2,2))
for (i in 1:4) biplotSel(pcmp::sce300xx, pcmp::acc4vis2, which=i)
par(op)
}
