% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\name{defaultProjectors}
\alias{defaultProjectors}
\title{schematize a sequence of projection methods}
\usage{
defaultProjectors()
}
\value{
a list with two components, 'projectors'
and 'retrievers'.  'projectors' is a named list of
 functions with parameters 'x' and 'ncomp',
where the intended usage has 'x' inheriting from SingleCellExperiment, 
and 'ncomp' numeric defining the number of projection components
to be retained.  'retrievers' is a named list of functions
with parameter 'x', which extracts the ncomp x nsamp
projection for each method.
}
\description{
schematize a sequence of projection methods
}
\note{
The default sequence is irlba::irlba, umap::umap,
Rtsne::Rtsne
}
\examples{
defaultProjectors
}
