% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\name{pcmpApp}
\alias{pcmpApp}
\title{Define a shiny app for comparing projections}
\usage{
pcmpApp(sce)
}
\arguments{
\item{sce}{SingleCellExperiment instance with reducedDims populated
with various candidates for projection from assay data --
currently ignored!  Just uses pcmp::sce300xx}
}
\description{
Define a shiny app for comparing projections
}
\note{
currently assuming that names(reducedDims) are 'PC', 'UM', 'TS'
for projections generated by
irlba::irlba, umap::umap, and Rtsne::Rtsne respectively
}
