% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\name{pcmpApp}
\alias{pcmpApp}
\title{Define a shiny app for comparing projections}
\usage{
pcmpApp(sce)
}
\arguments{
\item{sce}{SingleCellExperiment instance with reducedDims populated
with various candidates for projection from assay data --}
}
\value{
will return an instance of PcmpSels
}
\description{
Define a shiny app for comparing projections
}
\note{
In sce300xx, names(reducedDims) are 'PC', 'UM', 'TS'
for projections generated by
irlba::irlba, umap::umap, and Rtsne::Rtsne respectively.  Function
will delete .pcmpTab, .pcmpSelNum, .pcmpSelCells from .GlobalEnv
if found, so that it can assemble information on selections.
Note that tabulation of DE genes can be very slow if
the input SingleCellExperiment has more than 10k rows.
}
\examples{
if (interactive()) {
 message("be sure to use example(pcmpApp, ask=FALSE)")
 lit = pcmp::sce300xx
 sds = rowSds(assay(lit))
 top1k = order(sds,decreasing=TRUE)[1:1000]
 lit = lit[top1k,]
 chk = pcmpApp(lit) # fires up browser, if ask != FALSE you must
 # assent to each update 
 try(chk)  # may error if no selections
 try(head(geneTable(chk),3))  # likewise
}
}
