% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\name{pcmpApp}
\alias{pcmpApp}
\title{Define a shiny app for comparing projections}
\usage{
pcmpApp(sce)
}
\arguments{
\item{sce}{SingleCellExperiment instance with reducedDims populated
with various candidates for projection from assay data --}
}
\value{
will return an instance of PcmpSels
}
\description{
Define a shiny app for comparing projections
}
\note{
In sce300xx, names(reducedDims) are 'PC', 'UM', 'TS'
for projections generated by
irlba::irlba, umap::umap, and Rtsne::Rtsne respectively.  Function
will delete .pcmpTab, .pcmpSelNum, .pcmpSelCells from .GlobalEnv
if found, so that it can assemble information on selections.
}
